backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
sentence-transformers
python-dotenv
pydantic
langchain-google-genai
numpy